// lib/types.ts
export type RestrictedItem = {
  id: number;
  item_code: string;
  category: string;
  name: string;
  description?: string | null;

  is_serial_tracked: boolean;
  unit_name: string;

  quantity_on_hand: number;
  min_quantity?: number | null;

  serial_total?: number | null;
  serial_in_stock?: number | null;

  make_model?: string | null;
  caliber?: string | null;

  storage_location?: string | null;

  requires_maintenance: boolean;
  requires_cleaning: boolean;

  status: string;
  created_at: string;
  updated_at?: string | null;
};